+++ attributes/default.rb
@@ -1,1 +1,3 @@
 # Default Attributes For Appsindo Cookbook
+
+default['phpredis']['revision'] = 'master'