+++ recipes/nodejs.rb
@@ -13,6 +13,6 @@ include_recipe "nodejs"
 include_recipe "npm"
 
 # Install npm modules
-%w{ grunt-cli bower forever }.each do |a_package|
+%w{ grunt-cli bower forever yo }.each do |a_package|
   npm_package a_package
 end