+++ CHANGELOG.md
@@ -3,5 +3,5 @@ appsindo CHANGELOG
 
 0.1.0
 -----
-- [azureru] - Initial release of appsindo
+- [azureru] - Initial release of appsindo standard recipe